+++ src/main/java/com/samaxes/maven/minify/common/UglifyJS3Compiler.java
@@ -12,15 +12,15 @@ import org.apache.commons.io.IOUtils;
 
 /**
  * UglifyJS 3 compiler.
- *
+ * 
  * @author patchpump
  */
 public class UglifyJS3Compiler extends AbstractJavaScriptCompiler {
 
 	public UglifyJS3Compiler() throws IOException, ScriptException {
-		super("/META-INF/uglifyjs/uglifyjs-min-3.9.4.js");
+		super("/META-INF/uglifyjs/uglifyjs-min-3.4.9.js");
 	}
-
+	
 	@Override
 	public String compile(InputStreamReader reader) throws IOException, ScriptException, NoSuchMethodException {
 		String source = IOUtils.toString(reader);