+++ src/main/java/com/samaxes/maven/minify/plugin/ProcessJSFilesTask.java
@@ -34,7 +34,6 @@ import java.util.Map.Entry;
 
 import org.mozilla.javascript.EvaluatorException;
 
-import com.google.common.collect.Lists;
 import com.google.javascript.jscomp.CheckLevel;
 import com.google.javascript.jscomp.CommandLineRunner;
 import com.google.javascript.jscomp.Compiler;
@@ -43,6 +42,7 @@ import com.google.javascript.jscomp.DiagnosticGroup;
 import com.google.javascript.jscomp.DiagnosticGroups;
 import com.google.javascript.jscomp.SourceFile;
 import com.google.javascript.jscomp.SourceMap;
+import com.google.javascript.jscomp.jarjar.com.google.common.collect.Lists;
 import com.samaxes.maven.minify.common.JavaScriptErrorReporter;
 import com.samaxes.maven.minify.common.UglifyJS2Compiler;
 import com.samaxes.maven.minify.common.UglifyJS3Compiler;
@@ -57,7 +57,7 @@ public class ProcessJSFilesTask extends ProcessFilesTask {
 	 * Task constructor.
 	 *
 	 * @param opt task options
-	 * @throws FileNotFoundException
+	 * @throws FileNotFoundException 
 	 */
 	public ProcessJSFilesTask(final TaskOptions opt) throws FileNotFoundException {
 		super(opt);
@@ -111,7 +111,7 @@ public class ProcessJSFilesTask extends ProcessFilesTask {
 					options.setSourceMapOutputPath(sourceMapResult.getPath());
 				}
 
-				SourceFile input = SourceFile.fromInputStream(mergedFile.getName(), in);
+				SourceFile input = SourceFile.builder().withContent(in).withOriginalPath(mergedFile.getName()).build();
 				List<SourceFile> externs = opt.closureConfig.getExterns();
 				if (opt.closureConfig.getUseDefaultExterns()) {
 					externs.addAll(CommandLineRunner.getDefaultExterns());