+++ CHANGELOG.md
@@ -1,9 +1,18 @@
 # Minify Maven Plugin
 
+## 2.8.0-PATCHPUMP-R7
+
+* Add Nashorn 15.4 for JDK 17
+* Update Google Closure Compiler to v20230502
+
 ## 2.8.0-PATCHPUMP-R6
 
 * Update Google Closure Compiler to v20200517
 * Update UglifyJS 3 to v3.9.4
+
+## 2.8.0-PATCHPUMP-R5
+
+* Fix gzip run before source file is closed
 
 ## 2.8.0-PATCHPUMP-R4